PIA/requirements.txt
cryptography==35.0.0
lxml==4.6.3
requests==2.26.0
pyhunter==1.7
beautifulsoup4==4.10.0
Pillow==8.4.0
